Refactor import paths for pay-slip and salary-certificate forms
