+++ src/components/request-management/request-table-management.tsx
@@ -20,6 +20,7 @@ import {
 import { Badge } from '@/components/ui/badge';
 import { Calendar, FileText, DollarSign } from 'lucide-react';
 import { RequestDetails } from '@/types';
+import { useGetAllRequest } from '@/hooks';
 
 // Helper functions
 const formatDate = (dateString: string) =>
@@ -43,527 +44,14 @@ const getRequestState = (stateId: number) => {
   return states[stateId as keyof typeof states] || 'Desconocido';
 };
 
-interface RequestApproval {
-  id: number;
-  approverId: string | null;
-  requesterId: string;
-  processNumber: number;
-  RequestId: number;
-  observation: string | null;
-  approved: boolean | null;
-  current: boolean;
-  ApprovedDate: string | null;
-}
-
-interface RequestVacation {
-  id: number;
-  daysRequested: number;
-  departureDate: string;
-  entryDate: string;
-  RequestId: number;
-}
-
-export interface Request {
-  id: number;
-  date: string;
-  RequestStateId: number;
-  RequestTypeId: number;
-  EmployeeId: string;
-  RequestApprovals: RequestApproval[];
-  RequestVacation: RequestVacation | null;
-  RequestSalaryCertificate: null;
-  RequestPaymentConfirmation: null;
-};
-
-
-// Mock data
-const mockData: Request[] = [
-  {
-    id: 1,
-    date: '2024-10-07T01:45:26.000Z',
-    RequestStateId: 1,
-    RequestTypeId: 1,
-    EmployeeId: '504510677',
-    RequestApprovals: [
-      {
-        id: 1,
-        approverId: null,
-        requesterId: '504510677',
-        processNumber: 1,
-        RequestId: 1,
-        observation:
-          'La solicitud fue aprobada automáticamente en el proceso 1 ya que el solicitante es el jefe del departamento o no tiene un jefe de departamento.',
-        approved: true,
-        current: false,
-        ApprovedDate: '2024-10-07T01:45:26.000Z',
-      },
-      {
-        id: 2,
-        approverId: '504950876',
-        requesterId: '504510677',
-        processNumber: 2,
-        RequestId: 1,
-        observation: null,
-        approved: null,
-        current: true,
-        ApprovedDate: null,
-      },
-      {
-        id: 3,
-        approverId: '504420108',
-        requesterId: '504510677',
-        processNumber: 3,
-        RequestId: 1,
-        observation: null,
-        approved: null,
-        current: false,
-        ApprovedDate: null,
-      },
-    ],
-    RequestVacation: {
-      id: 1,
-      daysRequested: 7,
-      departureDate: '2024-10-06T06:00:00.000Z',
-      entryDate: '2024-10-13T06:00:00.000Z',
-      RequestId: 1,
-    },
-    RequestSalaryCertificate: null,
-    RequestPaymentConfirmation: null,
-  },
-  // 10 more vacation requests
-  {
-    id: 2,
-    date: '2024-10-08T09:30:00.000Z',
-    RequestStateId: 1,
-    RequestTypeId: 1,
-    EmployeeId: '504510678',
-    RequestApprovals: [
-      {
-        id: 4,
-        approverId: null,
-        requesterId: '504510678',
-        processNumber: 1,
-        RequestId: 2,
-        observation:
-          'La solicitud fue aprobada automáticamente en el proceso 1.',
-        approved: true,
-        current: false,
-        ApprovedDate: '2024-10-08T09:31:00.000Z',
-      },
-      {
-        id: 5,
-        approverId: '504950877',
-        requesterId: '504510678',
-        processNumber: 2,
-        RequestId: 2,
-        observation: null,
-        approved: null,
-        current: true,
-        ApprovedDate: null,
-      },
-    ],
-    RequestVacation: {
-      id: 2,
-      daysRequested: 5,
-      departureDate: '2024-11-01T06:00:00.000Z',
-      entryDate: '2024-11-06T06:00:00.000Z',
-      RequestId: 2,
-    },
-    RequestSalaryCertificate: null,
-    RequestPaymentConfirmation: null,
-  },
-  {
-    id: 3,
-    date: '2024-10-09T11:15:00.000Z',
-    RequestStateId: 2,
-    RequestTypeId: 1,
-    EmployeeId: '504510679',
-    RequestApprovals: [
-      {
-        id: 6,
-        approverId: null,
-        requesterId: '504510679',
-        processNumber: 1,
-        RequestId: 3,
-        observation:
-          'La solicitud fue aprobada automáticamente en el proceso 1.',
-        approved: true,
-        current: false,
-        ApprovedDate: '2024-10-09T11:16:00.000Z',
-      },
-      {
-        id: 7,
-        approverId: '504950878',
-        requesterId: '504510679',
-        processNumber: 2,
-        RequestId: 3,
-        observation: null,
-        approved: true,
-        current: false,
-        ApprovedDate: '2024-10-10T10:00:00.000Z',
-      },
-      {
-        id: 8,
-        approverId: '504420109',
-        requesterId: '504510679',
-        processNumber: 3,
-        RequestId: 3,
-        observation: null,
-        approved: null,
-        current: true,
-        ApprovedDate: null,
-      },
-    ],
-    RequestVacation: {
-      id: 3,
-      daysRequested: 10,
-      departureDate: '2024-11-15T06:00:00.000Z',
-      entryDate: '2024-11-25T06:00:00.000Z',
-      RequestId: 3,
-    },
-    RequestSalaryCertificate: null,
-    RequestPaymentConfirmation: null,
-  },
-  {
-    id: 4,
-    date: '2024-10-10T14:45:00.000Z',
-    RequestStateId: 3,
-    RequestTypeId: 1,
-    EmployeeId: '504510680',
-    RequestApprovals: [
-      {
-        id: 9,
-        approverId: null,
-        requesterId: '504510680',
-        processNumber: 1,
-        RequestId: 4,
-        observation:
-          'La solicitud fue rechazada debido a falta de días disponibles.',
-        approved: false,
-        current: false,
-        ApprovedDate: '2024-10-10T14:46:00.000Z',
-      },
-    ],
-    RequestVacation: {
-      id: 4,
-      daysRequested: 15,
-      departureDate: '2024-12-01T06:00:00.000Z',
-      entryDate: '2024-12-16T06:00:00.000Z',
-      RequestId: 4,
-    },
-    RequestSalaryCertificate: null,
-    RequestPaymentConfirmation: null,
-  },
-  {
-    id: 5,
-    date: '2024-10-11T16:00:00.000Z',
-    RequestStateId: 1,
-    RequestTypeId: 1,
-    EmployeeId: '504510681',
-    RequestApprovals: [
-      {
-        id: 10,
-        approverId: null,
-        requesterId: '504510681',
-        processNumber: 1,
-        RequestId: 5,
-        observation:
-          'La solicitud fue aprobada automáticamente en el proceso 1.',
-        approved: true,
-        current: false,
-        ApprovedDate: '2024-10-11T16:01:00.000Z',
-      },
-      {
-        id: 11,
-        approverId: '504950879',
-        requesterId: '504510681',
-        processNumber: 2,
-        RequestId: 5,
-        observation: null,
-        approved: null,
-        current: true,
-        ApprovedDate: null,
-      },
-    ],
-    RequestVacation: {
-      id: 5,
-      daysRequested: 7,
-      departureDate: '2024-11-20T06:00:00.000Z',
-      entryDate: '2024-11-27T06:00:00.000Z',
-      RequestId: 5,
-    },
-    RequestSalaryCertificate: null,
-    RequestPaymentConfirmation: null,
-  },
-  {
-    id: 6,
-    date: '2024-10-12T08:25:00.000Z',
-    RequestStateId: 2,
-    RequestTypeId: 1,
-    EmployeeId: '504510682',
-    RequestApprovals: [
-      {
-        id: 12,
-        approverId: null,
-        requesterId: '504510682',
-        processNumber: 1,
-        RequestId: 6,
-        observation:
-          'La solicitud fue aprobada automáticamente en el proceso 1.',
-        approved: true,
-        current: false,
-        ApprovedDate: '2024-10-12T08:26:00.000Z',
-      },
-      {
-        id: 13,
-        approverId: '504950880',
-        requesterId: '504510682',
-        processNumber: 2,
-        RequestId: 6,
-        observation: null,
-        approved: true,
-        current: false,
-        ApprovedDate: '2024-10-13T09:00:00.000Z',
-      },
-      {
-        id: 14,
-        approverId: '504420110',
-        requesterId: '504510682',
-        processNumber: 3,
-        RequestId: 6,
-        observation: null,
-        approved: null,
-        current: true,
-        ApprovedDate: null,
-      },
-    ],
-    RequestVacation: {
-      id: 6,
-      daysRequested: 12,
-      departureDate: '2024-12-05T06:00:00.000Z',
-      entryDate: '2024-12-17T06:00:00.000Z',
-      RequestId: 6,
-    },
-    RequestSalaryCertificate: null,
-    RequestPaymentConfirmation: null,
-  },
-  {
-    id: 7,
-    date: '2024-10-13T10:50:00.000Z',
-    RequestStateId: 1,
-    RequestTypeId: 1,
-    EmployeeId: '504510683',
-    RequestApprovals: [
-      {
-        id: 15,
-        approverId: null,
-        requesterId: '504510683',
-        processNumber: 1,
-        RequestId: 7,
-        observation:
-          'La solicitud fue aprobada automáticamente en el proceso 1.',
-        approved: true,
-        current: false,
-        ApprovedDate: '2024-10-13T10:51:00.000Z',
-      },
-      {
-        id: 16,
-        approverId: '504950881',
-        requesterId: '504510683',
-        processNumber: 2,
-        RequestId: 7,
-        observation: null,
-        approved: null,
-        current: true,
-        ApprovedDate: null,
-      },
-    ],
-    RequestVacation: {
-      id: 7,
-      daysRequested: 3,
-      departureDate: '2024-12-28T06:00:00.000Z',
-      entryDate: '2024-12-31T06:00:00.000Z',
-      RequestId: 7,
-    },
-    RequestSalaryCertificate: null,
-    RequestPaymentConfirmation: null,
-  },
-  {
-    id: 8,
-    date: '2024-10-14T13:15:00.000Z',
-    RequestStateId: 1,
-    RequestTypeId: 1,
-    EmployeeId: '504510684',
-    RequestApprovals: [
-      {
-        id: 17,
-        approverId: null,
-        requesterId: '504510684',
-        processNumber: 1,
-        RequestId: 8,
-        observation:
-          'La solicitud fue aprobada automáticamente en el proceso 1.',
-        approved: true,
-        current: false,
-        ApprovedDate: '2024-10-14T13:16:00.000Z',
-      },
-      {
-        id: 18,
-        approverId: '504950882',
-        requesterId: '504510684',
-        processNumber: 2,
-        RequestId: 8,
-        observation: null,
-        approved: null,
-        current: true,
-        ApprovedDate: null,
-      },
-    ],
-    RequestVacation: {
-      id: 8,
-      daysRequested: 14,
-      departureDate: '2025-01-05T06:00:00.000Z',
-      entryDate: '2025-01-19T06:00:00.000Z',
-      RequestId: 8,
-    },
-    RequestSalaryCertificate: null,
-    RequestPaymentConfirmation: null,
-  },
-  {
-    id: 9,
-    date: '2024-10-15T15:40:00.000Z',
-    RequestStateId: 2,
-    RequestTypeId: 1,
-    EmployeeId: '504510685',
-    RequestApprovals: [
-      {
-        id: 19,
-        approverId: null,
-        requesterId: '504510685',
-        processNumber: 1,
-        RequestId: 9,
-        observation:
-          'La solicitud fue aprobada automáticamente en el proceso 1.',
-        approved: true,
-        current: false,
-        ApprovedDate: '2024-10-15T15:41:00.000Z',
-      },
-      {
-        id: 20,
-        approverId: '504950883',
-        requesterId: '504510685',
-        processNumber: 2,
-        RequestId: 9,
-        observation: null,
-        approved: true,
-        current: false,
-        ApprovedDate: '2024-10-16T09:00:00.000Z',
-      },
-      {
-        id: 21,
-        approverId: '504420111',
-        requesterId: '504510685',
-        processNumber: 3,
-        RequestId: 9,
-        observation: null,
-        approved: null,
-        current: true,
-        ApprovedDate: null,
-      },
-    ],
-    RequestVacation: {
-      id: 9,
-      daysRequested: 8,
-      departureDate: '2025-01-20T06:00:00.000Z',
-      entryDate: '2025-01-28T06:00:00.000Z',
-      RequestId: 9,
-    },
-    RequestSalaryCertificate: null,
-    RequestPaymentConfirmation: null,
-  },
-  {
-    id: 10,
-    date: '2024-10-16T17:05:00.000Z',
-    RequestStateId: 3,
-    RequestTypeId: 1,
-    EmployeeId: '504510686',
-    RequestApprovals: [
-      {
-        id: 22,
-        approverId: null,
-        requesterId: '504510686',
-        processNumber: 1,
-        RequestId: 10,
-        observation:
-          'La solicitud fue rechazada debido a conflicto con proyectos en curso.',
-        approved: false,
-        current: false,
-        ApprovedDate: '2024-10-16T17:06:00.000Z',
-      },
-    ],
-    RequestVacation: {
-      id: 10,
-      daysRequested: 6,
-      departureDate: '2024-12-15T06:00:00.000Z',
-      entryDate: '2024-12-21T06:00:00.000Z',
-      RequestId: 10,
-    },
-    RequestSalaryCertificate: null,
-    RequestPaymentConfirmation: null,
-  },
-  {
-    id: 11,
-    date: '2024-10-17T19:30:00.000Z',
-    RequestStateId: 1,
-    RequestTypeId: 1,
-    EmployeeId: '504510687',
-    RequestApprovals: [
-      {
-        id: 23,
-        approverId: null,
-        requesterId: '504510687',
-        processNumber: 1,
-        RequestId: 11,
-        observation:
-          'La solicitud fue aprobada automáticamente en el proceso 1.',
-        approved: true,
-        current: false,
-        ApprovedDate: '2024-10-17T19:31:00.000Z',
-      },
-      {
-        id: 24,
-        approverId: '504950884',
-        requesterId: '504510687',
-        processNumber: 2,
-        RequestId: 11,
-        observation: null,
-        approved: null,
-        current: true,
-        ApprovedDate: null,
-      },
-    ],
-    RequestVacation: {
-      id: 11,
-      daysRequested: 4,
-      departureDate: '2024-11-25T06:00:00.000Z',
-      entryDate: '2024-11-29T06:00:00.000Z',
-      RequestId: 11,
-    },
-    RequestSalaryCertificate: null,
-    RequestPaymentConfirmation: null,
-  },
-];
-
-
-
 const RequestTable = ({
   requests,
   onRowClick,
 }: {
-  requests: Request[];
-  onRowClick: (request: Request) => void;
+  requests: RequestDetails[];
+  onRowClick: (request: RequestDetails) => void;
 }) => (
-  <div className='border rounded shadow'>
+  <div className="border rounded shadow">
     <Table>
       <TableHeader>
         <TableRow>
@@ -610,7 +98,7 @@ const RequestModal = ({
   isOpen,
   onClose,
 }: {
-  request: Request | null;
+  request: RequestDetails | null;
   isOpen: boolean;
   onClose: () => void;
 }) => {
@@ -715,9 +203,10 @@ const RequestModal = ({
 };
 
 export default function RequestTableManagement() {
-  const [selectedRequest, setSelectedRequest] = useState<Request | null>(null);
+  const [selectedRequest, setSelectedRequest] = useState<RequestDetails | null>(null);
+  const { allRequests } = useGetAllRequest();
 
-  const handleRowClick = (request: Request) => {
+  const handleRowClick = (request: RequestDetails) => {
     setSelectedRequest(request);
   };
 
@@ -728,7 +217,7 @@ export default function RequestTableManagement() {
   return (
     <div className="container mx-auto py-10">
       <h1 className="text-2xl font-bold mb-4">Gestión de Solicitudes</h1>
-      <RequestTable requests={mockData} onRowClick={handleRowClick} />
+      <RequestTable requests={allRequests || []} onRowClick={handleRowClick} />
       <RequestModal
         request={selectedRequest}
         isOpen={!!selectedRequest}
